@@ -3,72 +3,76 @@
     "version": "0.0.0",
     "private": true,
     "scripts": {
-        "dev": "next dev",
         "build": "next build",
-        "start": "next start",
-        "lint": "next lint",
         "clean": "rm -rf node_modules && rm -rf .next",
-        "lint:fix": "eslint . --ext .ts,.js,.tsx,.jsx --fix && pnpm run format",
-        "generate": "prisma generate && prisma format",
-        "pre-commit": "lint-staged",
+        "db-deploy": "pnpm prisma migrate deploy",
+        "db-migrate": "pnpm prisma migrate dev",
+        "dev": "next dev",
         "format": "prettier --write \"**/*.{ts,tsx,md}\"",
-        "postinstall": "pnpm run generate"
+        "generate": "prisma generate && prisma format",
+        "postinstall": "pnpm run generate",
+        "lint": "next lint",
+        "lint:fix": "eslint . --ext .ts,.js,.tsx,.jsx --fix && pnpm run format",
+        "pre-commit": "lint-staged",
+        "start": "next start"
     },
     "dependencies": {
         "@formatjs/intl-localematcher": "^0.5.4",
-        "@formkit/auto-animate": "^0.8.1",
-        "@hookform/resolvers": "^3.3.4",
-        "@prisma/client": "5.11.0",
-        "@radix-ui/react-dropdown-menu": "^2.0.6",
-        "@radix-ui/react-label": "^2.0.2",
-        "@radix-ui/react-slot": "^1.0.2",
-        "@radix-ui/react-tooltip": "^1.0.7",
+        "@formkit/auto-animate": "^0.8.2",
+        "@hookform/resolvers": "^3.6.0",
+        "@prisma/client": "5.15.1",
+        "@radix-ui/react-dropdown-menu": "^2.1.0",
+        "@radix-ui/react-label": "^2.1.0",
+        "@radix-ui/react-slot": "^1.1.0",
+        "babel-plugin-react-compiler": "0.0.0-experimental-938cd9a-20240601",
         "class-variance-authority": "^0.7.0",
-        "clsx": "^2.1.0",
-        "dayjs": "^1.11.10",
+        "clsx": "^2.1.1",
+        "dayjs": "^1.11.11",
         "device-uuid": "^1.0.4",
-        "framer-motion": "^11.0.14",
-        "lucide-react": "^0.358.0",
+        "framer-motion": "^11.2.11",
+        "lucide-react": "^0.395.0",
         "negotiator": "^0.6.3",
-        "next": "14.1.3",
-        "postcss": "8.4.36",
-        "prisma": "^5.11.0",
-        "react": "18.2.0",
-        "react-dom": "18.2.0",
-        "react-hook-form": "^7.51.1",
+        "next": "15.0.0-rc.0",
+        "prisma": "^5.15.1",
+        "react": "19.0.0-rc-e684ca66ab-20240619",
+        "react-dom": "19.0.0-rc-e684ca66ab-20240619",
+        "react-hook-form": "^7.52.0",
+        "react-textarea-autosize": "^8.5.3",
         "server-only": "^0.0.1",
-        "tailwind-merge": "^2.2.2",
-        "tailwindcss": "3.4.1",
+        "tailwind-merge": "^2.3.0",
+        "tailwindcss": "3.4.4",
         "tailwindcss-animate": "^1.0.7",
-        "vaul": "^0.9.0",
-        "zod": "^3.22.4",
+        "vaul": "^0.9.1",
+        "zod": "^3.23.8",
         "zustand": "^4.5.2"
+    },
+    "devDependencies": {
+        "@ianvs/prettier-plugin-sort-imports": "^4.2.1",
+        "@types/eslint": "^8.56.10",
+        "@types/negotiator": "^0.6.3",
+        "@types/node": "20.14.6",
+        "@types/react": "18.3.3",
+        "@types/react-dom": "18.3.0",
+        "@typescript-eslint/eslint-plugin": "^7.13.1",
+        "@typescript-eslint/parser": "^7.13.1",
+        "autoprefixer": "10.4.19",
+        "eslint": "8.57.0",
+        "eslint-config-next": "14.2.4",
+        "husky": "^9.0.11",
+        "lint-staged": "^15.2.7",
+        "postcss": "8.4.38",
+        "postcss-import": "^16.1.0",
+        "postcss-preset-env": "^9.5.14",
+        "prettier": "^3.3.2",
+        "prettier-plugin-packagejson": "^2.5.0",
+        "prettier-plugin-tailwindcss": "^0.6.5",
+        "pretty-quick": "^4.0.0",
+        "typescript": "5.4.5"
     },
     "packageManager": "pnpm@7.18.1",
     "engines": {
         "node": ">=18.x",
         "npm": ">=9.0.0"
     },
-    "devDependencies": {
-        "@ianvs/prettier-plugin-sort-imports": "^4.2.1",
-        "@types/eslint": "^8.56.5",
-        "@types/negotiator": "^0.6.3",
-        "@types/node": "20.11.28",
-        "@types/react": "18.2.66",
-        "@types/react-dom": "18.2.22",
-        "@typescript-eslint/eslint-plugin": "^7.2.0",
-        "@typescript-eslint/parser": "^7.2.0",
-        "autoprefixer": "10.4.18",
-        "eslint": "8.57.0",
-        "eslint-config-next": "14.1.3",
-        "husky": "^9.0.11",
-        "lint-staged": "^15.2.2",
-        "postcss-import": "^16.0.1",
-        "postcss-preset-env": "^9.5.2",
-        "prettier": "^3.2.5",
-        "prettier-plugin-packagejson": "^2.4.12",
-        "prettier-plugin-tailwindcss": "^0.5.12",
-        "pretty-quick": "^4.0.0",
-        "typescript": "5.4.2"
-    }
+    "engineStrict": true
 }
