@@ -41,12 +41,11 @@ const useFormField = () => {
     const itemContext = React.useContext(FormItemContext);
     const { getFieldState, formState } = useFormContext();
 
-    const fieldState = getFieldState(fieldContext.name, formState);
-
     if (!fieldContext) {
         throw new Error("useFormField should be used within <FormField>");
     }
 
+    const fieldState = getFieldState(fieldContext.name, formState);
     const { id } = itemContext;
 
     return {
@@ -65,45 +64,34 @@ type FormItemContextValue = {
 
 const FormItemContext = React.createContext<FormItemContextValue>({} as FormItemContextValue);
 
-const FormItem = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
-    ({ className, ...props }, ref) => {
-        const id = React.useId();
+const FormItem = ({ className, ...props }: React.ComponentPropsWithoutRef<"div">) => {
+    const id = React.useId();
+    return (
+        <FormItemContext.Provider value={{ id }}>
+            <div className={cn("space-y-2", className)} {...props} />
+        </FormItemContext.Provider>
+    );
+};
 
-        return (
-            <FormItemContext.Provider value={{ id }}>
-                <div ref={ref} className={cn("space-y-2", className)} {...props} />
-            </FormItemContext.Provider>
-        );
-    }
-);
-FormItem.displayName = "FormItem";
-
-const FormLabel = React.forwardRef<
-    React.ElementRef<typeof LabelPrimitive.Root>,
-    React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root>
->(({ className, ...props }, ref) => {
+const FormLabel = ({
+    className,
+    ...props
+}: React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root>) => {
     const { error, formItemId } = useFormField();
 
     return (
         <Label
-            ref={ref}
             className={cn(error && "text-destructive", className)}
             htmlFor={formItemId}
             {...props}
         />
     );
-});
-FormLabel.displayName = "FormLabel";
+};
 
-const FormControl = React.forwardRef<
-    React.ElementRef<typeof Slot>,
-    React.ComponentPropsWithoutRef<typeof Slot>
->(({ ...props }, ref) => {
+const FormControl = (props: React.ComponentPropsWithoutRef<typeof Slot>) => {
     const { error, formItemId, formDescriptionId, formMessageId } = useFormField();
-
     return (
         <Slot
-            ref={ref}
             id={formItemId}
             aria-describedby={
                 !error ? `${formDescriptionId}` : `${formDescriptionId} ${formMessageId}`
@@ -112,48 +100,38 @@ const FormControl = React.forwardRef<
             {...props}
         />
     );
-});
-FormControl.displayName = "FormControl";
+};
 
-const FormDescription = React.forwardRef<
-    HTMLParagraphElement,
-    React.HTMLAttributes<HTMLParagraphElement>
->(({ className, ...props }, ref) => {
+const FormDescription = ({ className, ...props }: React.HTMLAttributes<HTMLParagraphElement>) => {
     const { formDescriptionId } = useFormField();
 
     return (
         <p
-            ref={ref}
             id={formDescriptionId}
             className={cn("text-sm text-muted-foreground", className)}
             {...props}
         />
     );
-});
-FormDescription.displayName = "FormDescription";
+};
 
-const FormMessage = React.forwardRef<
-    HTMLParagraphElement,
-    React.HTMLAttributes<HTMLParagraphElement>
->(({ className, children, ...props }, ref) => {
+const FormMessage = ({
+    className,
+    children,
+    ...props
+}: React.HTMLAttributes<HTMLParagraphElement>) => {
     const { error, formMessageId } = useFormField();
     const body = error ? String(error?.message) : children;
-
-    if (!body) {
-        return null;
-    }
+    if (!body) return null;
 
     return (
         <p
-            ref={ref}
             id={formMessageId}
             className={cn("text-sm font-medium text-destructive", className)}
             {...props}>
             {body}
         </p>
     );
-});
-FormMessage.displayName = "FormMessage";
+};
 
 export {
     useFormField,
