@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const config = {};
+const config = {
+    experimental: {
+        reactCompiler: true,
+    },
+};
 
 export default config;
